removed uint8_t to fix python binding problem

+++ src/openms/include/OpenMS/CHEMISTRY/ISOTOPEDISTRIBUTION/CoarseIsotopePatternGenerator.h
@@ -173,7 +173,7 @@ namespace OpenMS
        @param average_weight: m/z of monoisotopic peak (with charge = 1) to approximate the distribution of intensities for
        @param num_peaks: How many peaks should be generated (independent of this->max_isotope)
     */
-    static IsotopeDistribution approximateFromPeptideWeight(double mass, UInt num_peaks = 20, uint8_t charge = 1);
+    static IsotopeDistribution approximateFromPeptideWeight(double mass, UInt num_peaks = 20, UInt charge = 1);
 
     /**
        @brief roughly approximate intensity distribution of peptidic isotope patterns from monoisotopic weight using Poisson distribution.
